refactor: Systemic decoupling of STT pipeline into specialized modules
